views refactored
