--- indieauth/_models.py
@@ -1,42 +1,0 @@
-import base64
-from datetime import timedelta
-import hashlib
-import random
-import socket
-import string
-import ipaddress
-import re
-from urllib.parse import urljoin, urlsplit
-
-from bs4 import BeautifulSoup
-from django.db import models
-from django.utils import timezone
-from django.contrib.auth.models import User
-import mf2py
-from requests import Response, request
-
-from orangegnome import settings
-from profiles.models import Profile
-
-# Create your models here.
-
-
-
-
-    
-
-        
-
-
-
-
-    
-
-    
-
-    
-
-    
-
-    
-
