+++ indieauth/viewmodels.py
@@ -1,6 +1,6 @@
 from urllib.parse import urlsplit
 from django.http import QueryDict
-from indieauth.models import AuthCode, ClientMetadata
+from indieauth.models import ClientMetadata
 
 class AuthRequestVM():
     ACCEPT:str = "ACCEPT"
@@ -78,13 +78,3 @@ class AuthSubmissionVM():
 
         return True 
         
-    def create_auth_code(self, user_id):
-        code = AuthCode.create()
-        code.client_id = self.values.get("client_id")
-        code.redirect_uri = self.values.get("redirect_uri")
-        code.user_id = user_id
-        code.code_challenge_method = self.values.get("code_challenge_method")
-        code.code_challenge = self.values.get("code_challenge")
-        code.scope = self.values.get("scope", "")
-        
-        return code
