+++ indieauth/views/__init__.py
@@ -1,0 +1,6 @@
+from .metadata import MetadataView
+from .token import TokenView
+from .auth import AuthView
+from .introspect import IntrospectView
+from .revoke import RevokeView
+from .userinfo import UserInfoView
